add replications, add the steps to README

+++ numbers_and_pvalues/ASE_replications/LCL_replication/merge_with_LCL_ASE.py
@@ -55,7 +55,7 @@ for f in glob.glob(data_dir+'/counts_summed*txt'):
             elif int(line[3]) == 0:
                 continue
             else:
-                logFC = math.log(int(line[4])/int(line[3]))
+                logFC = math.log(float(line[4])/float(line[3]))
